Simplification of AutoSpanGridLayoutManager
