+++ app/src/main/java/com/infomaniak/drive/views/AutoSpanGridLayoutManager.kt
@@ -1,0 +1,53 @@
+/*
+ * Infomaniak kDrive - Android
+ * Copyright (C) 2022 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.drive.views
+
+import android.content.Context
+import android.util.AttributeSet
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.infomaniak.lib.core.utils.toPx
+import kotlin.math.max
+
+class AutoSpanGridLayoutManager @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyle: Int = 0,
+    defStyleRes: Int = 0
+) : GridLayoutManager(context, attrs, defStyle, defStyleRes) {
+
+    private val columnWidth = 56.toPx()
+
+    private var previousWidth = 0
+    private var previousHeight = 0
+
+    override fun onLayoutChildren(recycler: RecyclerView.Recycler, state: RecyclerView.State) {
+        if (width > 0 && height > 0 && (previousWidth != width || previousHeight != height)) {
+            val totalSpace = if (orientation == VERTICAL) {
+                width - paddingRight - paddingLeft
+            } else {
+                height - paddingTop - paddingBottom
+            }
+            val spanCount = max(1, totalSpace / columnWidth)
+            setSpanCount(spanCount)
+        }
+        previousWidth = width
+        previousHeight = height
+        super.onLayoutChildren(recycler, state)
+    }
+}
