+++ app/src/main/res/layout/fragment_bottom_sheet_color_folder.xml
@@ -71,7 +71,7 @@
         android:layout_marginTop="@dimen/marginStandardSmall"
         android:overScrollMode="never"
         android:paddingHorizontal="@dimen/marginStandardMedium"
+        app:layoutManager="com.infomaniak.drive.views.AutoSpanGridLayoutManager"
         tools:itemCount="17"
-        tools:listitem="@layout/view_color_round_button"
-        tools:spanCount="6" />
+        tools:listitem="@layout/view_color_round_button" />
 </LinearLayout>
