+++ app/src/main/res/layout/fragment_create_or_edit_category.xml
@@ -122,6 +122,7 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/chooseCategory"
                 tools:itemCount="23"
+                tools:layoutManager="com.infomaniak.drive.views.AutoSpanGridLayoutManager"
                 tools:listitem="@layout/view_color_round_button"
                 tools:spanCount="7" />
